+++ src/juscraper/courts/tjrn/schemas.py
@@ -28,7 +28,7 @@ class InputCJSGTJRN(SearchBase, DataJulgamentoMixin):
     BACKEND_DATE_FORMAT: ClassVar[str] = "%Y-%m-%d"
 
     numero_processo: str = ""
-    id_classe_judicial: str = ""
+    id_classe: str = ""
     id_orgao_julgador: str = ""
     id_relator: str = ""
     id_colegiado: str = ""